+++ spiffworkflow-backend/src/spiffworkflow_backend/routes/process_api_blueprint.py
@@ -1,18 +1,28 @@
+import json
+import uuid
 from typing import Any
 from uuid import UUID
 
 import flask.wrappers
+import sentry_sdk
 from flask import Blueprint
 from flask import current_app
 from flask import g
 from flask import jsonify
 from flask import make_response
 from flask.wrappers import Response
+from SpiffWorkflow.task import Task as SpiffTask  # type: ignore
+from SpiffWorkflow.util.task import TaskState  # type: ignore
 from sqlalchemy import and_
 from sqlalchemy import or_
 
+from spiffworkflow_backend.background_processing.celery_tasks.process_instance_task_producer import (
+    queue_enabled_for_process_model,
+)
+from spiffworkflow_backend.data_migrations.process_instance_migrator import ProcessInstanceMigrator
 from spiffworkflow_backend.exceptions.api_error import ApiError
 from spiffworkflow_backend.exceptions.process_entity_not_found_error import ProcessEntityNotFoundError
+from spiffworkflow_backend.models.db import db
 from spiffworkflow_backend.models.human_task import HumanTaskModel
 from spiffworkflow_backend.models.human_task_user import HumanTaskUserModel
 from spiffworkflow_backend.models.principal import PrincipalModel
@@ -21,12 +31,19 @@ from spiffworkflow_backend.models.process_instance_file_data import ProcessInsta
 from spiffworkflow_backend.models.process_model import ProcessModelInfo
 from spiffworkflow_backend.models.reference_cache import ReferenceCacheModel
 from spiffworkflow_backend.models.reference_cache import ReferenceSchema
+from spiffworkflow_backend.models.task import TaskModel  # noqa: F401
 from spiffworkflow_backend.services.authentication_service import AuthenticationService  # noqa: F401
 from spiffworkflow_backend.services.authorization_service import AuthorizationService
+from spiffworkflow_backend.services.git_service import GitCommandError
 from spiffworkflow_backend.services.git_service import GitService
+from spiffworkflow_backend.services.jinja_service import JinjaService
 from spiffworkflow_backend.services.process_caller_service import ProcessCallerService
 from spiffworkflow_backend.services.process_instance_processor import ProcessInstanceProcessor
+from spiffworkflow_backend.services.process_instance_queue_service import ProcessInstanceQueueService
+from spiffworkflow_backend.services.process_instance_service import ProcessInstanceService
 from spiffworkflow_backend.services.process_model_service import ProcessModelService
+from spiffworkflow_backend.services.task_service import TaskModelError
+from spiffworkflow_backend.services.task_service import TaskService
 
 process_api_blueprint = Blueprint("process_api", __name__)
 
@@ -357,3 +374,189 @@ def _get_process_model_for_instantiation(
             status_code=400,
         )
     return process_model
+
+
+def _prepare_form_data(
+    form_file: str, process_model: ProcessModelInfo, task_model: TaskModel | None = None, revision: str | None = None
+) -> dict:
+    try:
+        form_contents = GitService.get_file_contents_for_revision_if_git_revision(
+            process_model=process_model,
+            revision=revision,
+            file_name=form_file,
+        )
+    except GitCommandError as exception:
+        raise (
+            ApiError(
+                error_code="git_error_loading_form",
+                message=(
+                    f"Could not load form schema from: {form_file}. Was git history rewritten such that revision"
+                    f" '{revision}' no longer exists? Error was: {str(exception)}"
+                ),
+                status_code=400,
+            )
+        ) from exception
+
+    if task_model and task_model.data is not None:
+        try:
+            form_contents = JinjaService.render_jinja_template(form_contents, task=task_model)
+        except TaskModelError as wfe:
+            wfe.add_note(f"Error in Json Form File '{form_file}'")
+            api_error = ApiError.from_workflow_exception("instructions_error", str(wfe), exp=wfe)
+            api_error.file_name = form_file
+            raise api_error from wfe
+
+    try:
+        # form_contents is a str
+        hot_dict: dict = json.loads(form_contents)
+        return hot_dict
+    except Exception as exception:
+        raise (
+            ApiError(
+                error_code="error_loading_form",
+                message=f"Could not load form schema from: {form_file}. Error was: {str(exception)}",
+                status_code=400,
+            )
+        ) from exception
+
+
+def _task_submit_shared(
+    process_instance_id: int,
+    task_guid: str,
+    body: dict[str, Any],
+    execution_mode: str | None = None,
+) -> dict:
+    principal = _find_principal_or_raise()
+    process_instance = _find_process_instance_by_id_or_raise(process_instance_id)
+    if not process_instance.can_submit_task():
+        raise ApiError(
+            error_code="process_instance_not_runnable",
+            message=(
+                f"Process Instance ({process_instance.id}) has status "
+                f"{process_instance.status} which does not allow tasks to be submitted."
+            ),
+            status_code=400,
+        )
+
+    # we're dequeing twice in this function.
+    # tried to wrap the whole block in one dequeue, but that has the confusing side-effect that every exception
+    # in the block causes the process instance to go into an error state. for example, when
+    # AuthorizationService.assert_user_can_complete_task raises. this would have been solvable, but this seems simpler,
+    # and the cost is not huge given that this function is not the most common code path in the world.
+    with ProcessInstanceQueueService.dequeued(process_instance):
+        ProcessInstanceMigrator.run(process_instance)
+
+    processor = ProcessInstanceProcessor(
+        process_instance, workflow_completed_handler=ProcessInstanceService.schedule_next_process_model_cycle
+    )
+    spiff_task = _get_spiff_task_from_processor(task_guid, processor)
+    AuthorizationService.assert_user_can_complete_task(process_instance.id, str(spiff_task.id), principal.user)
+
+    if spiff_task.state != TaskState.READY:
+        raise (
+            ApiError(
+                error_code="invalid_state",
+                message="You may not update a task unless it is in the READY state.",
+                status_code=400,
+            )
+        )
+
+    human_task = _find_human_task_or_raise(
+        process_instance_id=process_instance_id,
+        task_guid=task_guid,
+        only_tasks_that_can_be_completed=True,
+    )
+
+    with sentry_sdk.start_span(op="task", description="complete_form_task"):
+        with ProcessInstanceQueueService.dequeued(process_instance):
+            ProcessInstanceService.complete_form_task(
+                processor=processor,
+                spiff_task=spiff_task,
+                data=body,
+                user=g.user,
+                human_task=human_task,
+                execution_mode=execution_mode,
+            )
+
+    # currently task_model has the potential to be None. This should be removable once
+    # we backfill the human_task table for task_guid and make that column not nullable
+    task_model: TaskModel | None = human_task.task_model
+    if task_model is None:
+        task_model = TaskModel.query.filter_by(guid=human_task.task_id).first()
+
+    # delete draft data when we submit a task to ensure cycling back to the task contains the
+    # most up-to-date data
+    task_draft_data = TaskService.task_draft_data_from_task_model(task_model)
+    if task_draft_data is not None:
+        db.session.delete(task_draft_data)
+        db.session.commit()
+
+    next_human_task_assigned_to_me = TaskService.next_human_task_for_user(process_instance_id, principal.user_id)
+    if next_human_task_assigned_to_me:
+        return {"next_human_task_assigned_to_me": HumanTaskModel.to_task(next_human_task_assigned_to_me)}
+
+    # a guest user completed a task, it has a guest_confirmation message to display to them,
+    # and there is nothing else for them to do
+    spiff_task_extensions = spiff_task.task_spec.extensions
+    if (
+        "allowGuest" in spiff_task_extensions
+        and spiff_task_extensions["allowGuest"] == "true"
+        and "guestConfirmation" in spiff_task.task_spec.extensions
+    ):
+        return {"guest_confirmation": spiff_task.task_spec.extensions["guestConfirmation"]}
+
+    if processor.next_task():
+        task = ProcessInstanceService.spiff_task_to_api_task(processor, processor.next_task())
+        task.process_model_uses_queued_execution = queue_enabled_for_process_model(process_instance)
+        return {"next_task": task}
+
+    # next_task always returns something, even if the instance is complete, so we never get here
+    return {
+        "ok": True,
+        "process_model_identifier": process_instance.process_model_identifier,
+        "process_instance_id": process_instance_id,
+    }
+
+
+def _find_human_task_or_raise(
+    process_instance_id: int,
+    task_guid: str,
+    only_tasks_that_can_be_completed: bool = False,
+) -> HumanTaskModel:
+    if only_tasks_that_can_be_completed:
+        human_task_query = HumanTaskModel.query.filter_by(
+            process_instance_id=process_instance_id,
+            task_id=task_guid,
+            completed=False,
+        )
+    else:
+        human_task_query = HumanTaskModel.query.filter_by(process_instance_id=process_instance_id, task_id=task_guid)
+
+    human_task: HumanTaskModel = human_task_query.first()
+    if human_task is None:
+        raise (
+            ApiError(
+                error_code="no_human_task",
+                message=f"Cannot find a task to complete for task id '{task_guid}' and process instance {process_instance_id}.",
+                status_code=500,
+            )
+        )
+    return human_task
+
+
+def _get_spiff_task_from_processor(
+    task_guid: str,
+    processor: ProcessInstanceProcessor,
+) -> SpiffTask:
+    task_uuid = uuid.UUID(task_guid)
+    spiff_task = processor.bpmn_process_instance.get_task_from_id(task_uuid)
+
+    if spiff_task is None:
+        raise (
+            ApiError(
+                error_code="empty_task",
+                message="Processor failed to obtain task.",
+                status_code=500,
+            )
+        )
+    return spiff_task
