+++ spiffworkflow-backend/src/spiffworkflow_backend/routes/public_controller.py
@@ -1,4 +1,3 @@
-import json
 from typing import Any
 
 import flask.wrappers
@@ -11,10 +10,12 @@ from spiffworkflow_backend.exceptions.api_error import ApiError
 from spiffworkflow_backend.models.process_instance import ProcessInstanceModel
 from spiffworkflow_backend.models.process_instance import ProcessInstanceStatus
 from spiffworkflow_backend.models.process_model import ProcessModelInfo
+from spiffworkflow_backend.models.task import TaskModel  # noqa: F401
+from spiffworkflow_backend.routes.process_api_blueprint import _prepare_form_data
+from spiffworkflow_backend.routes.process_api_blueprint import _task_submit_shared
 from spiffworkflow_backend.services.message_service import MessageService
 from spiffworkflow_backend.services.process_instance_processor import ProcessInstanceProcessor
 from spiffworkflow_backend.services.process_model_service import ProcessModelService
-from spiffworkflow_backend.services.spec_file_service import SpecFileService
 from spiffworkflow_backend.services.task_service import TaskService
 
 
@@ -47,17 +48,8 @@ def message_form_show(
         )
 
     process_model = ProcessModelService.get_process_model(message_triggerable_process_model.process_model_identifier)
-
-    response_body = {}
     extensions = matching_start_tasks[0].task_spec.extensions
-    if "properties" in extensions:
-        properties = extensions["properties"]
-        if "formJsonSchemaFilename" in properties:
-            form_schema_file_name = properties["formJsonSchemaFilename"]
-            response_body["form_schema"] = _get_json_contents_from_file(form_schema_file_name, process_model)
-        if "formUiSchemaFilename" in properties:
-            form_ui_schema_file_name = properties["formUiSchemaFilename"]
-            response_body["form_ui_schema"] = _get_json_contents_from_file(form_ui_schema_file_name, process_model)
+    response_body = _get_form_and_prepare_data(extensions=extensions, process_model=process_model)
 
     return make_response(jsonify(response_body), 200)
 
@@ -71,13 +63,86 @@ def message_form_submit(
     process_instance = ProcessInstanceModel.query.filter_by(id=receiver_message.process_instance_id).first()
     next_human_task_assigned_to_me = TaskService.next_human_task_for_user(process_instance.id, g.user.id)
 
+    next_form_contents = None
+    task_guid = None
+    if next_human_task_assigned_to_me:
+        task_guid = next_human_task_assigned_to_me.task_guid
+        process_model = ProcessModelService.get_process_model(process_instance.process_model_identifier)
+        next_form_contents = _get_form_and_prepare_data(
+            process_model=process_model, task_guid=next_human_task_assigned_to_me.task_guid, process_instance=process_instance
+        )
+
     response_json = {
-        "next_task": next_human_task_assigned_to_me,
+        "form": next_form_contents,
+        "task_guid": task_guid,
+        "process_instance_id": process_instance.id,
         "instructions": None,
     }
     return make_response(jsonify(response_json), 200)
 
 
-def _get_json_contents_from_file(file_name: str, process_model: ProcessModelInfo) -> dict:
-    contents = SpecFileService.get_data(process_model, file_name).decode("utf-8")
-    return dict(json.loads(contents))
+def form_submit(
+    process_instance_id: int,
+    task_guid: str,
+    body: dict[str, Any],
+    execution_mode: str | None = None,
+) -> flask.wrappers.Response:
+    response_item = _task_submit_shared(process_instance_id, task_guid, body, execution_mode=execution_mode)
+
+    next_form_contents = None
+    next_task_guid = None
+    if "next_human_task_assigned_to_me" in response_item:
+        next_human_task_assigned_to_me = response_item["next_human_task_assigned_to_me"]
+        process_instance = ProcessInstanceModel.query.filter_by(id=process_instance_id).first()
+        next_task_guid = next_human_task_assigned_to_me.task_guid
+        process_model = ProcessModelService.get_process_model(process_instance.process_model_identifier)
+        next_form_contents = _get_form_and_prepare_data(
+            process_model=process_model, task_guid=next_human_task_assigned_to_me.task_guid, process_instance=process_instance
+        )
+
+    response_json = {
+        "form": next_form_contents,
+        "task_guid": next_task_guid,
+        "process_instance_id": process_instance_id,
+        "instructions": None,
+    }
+    return make_response(jsonify(response_json), 200)
+
+
+def _get_form_and_prepare_data(
+    process_model: ProcessModelInfo,
+    extensions: dict | None = None,
+    task_guid: str | None = None,
+    process_instance: ProcessInstanceModel | None = None,
+) -> dict:
+    task_model = None
+    extension_list = extensions
+    if task_guid and process_instance:
+        task_model = TaskModel.query.filter_by(guid=task_guid, process_instance_id=process_instance.id).first()
+        task_model.data = task_model.json_data()
+        extension_list = TaskService.get_extensions_from_task_model(task_model)
+
+    revision = None
+    if process_instance:
+        revision = process_instance.bpmn_version_control_identifier
+
+    form_contents = {}
+    if extension_list and "properties" in extension_list:
+        properties = extension_list["properties"]
+        if "formJsonSchemaFilename" in properties:
+            form_schema_file_name = properties["formJsonSchemaFilename"]
+            form_contents["form_schema"] = _prepare_form_data(
+                form_file=form_schema_file_name,
+                task_model=task_model,
+                process_model=process_model,
+                revision=revision,
+            )
+        if "formUiSchemaFilename" in properties:
+            form_ui_schema_file_name = properties["formUiSchemaFilename"]
+            form_contents["form_ui_schema"] = _prepare_form_data(
+                form_file=form_ui_schema_file_name,
+                task_model=task_model,
+                process_model=process_model,
+                revision=revision,
+            )
+    return form_contents
